+++ new/keyvalue.h
@@ -16,6 +16,7 @@
 
 #include "mpi.h"
 #include "stdio.h"
+#include "stdint.h"
 
 namespace MAPREDUCE_NS {
 
@@ -74,7 +75,7 @@ class KeyValue {
     int exactsize;              // exact size of all data in page
     int alignsize;              // rounded-up exactsize with alignment
     int filesize;               // rounded-up alignsize for file I/O
-    int fileoffset;             // summed filesize of all previous pages
+    uint64_t fileoffset;        // summed filesize of all previous pages
   };
 
   Page *pages;                  // list of pages
